Add getters/setters for node `children`
Using these properties a tree can be added to or taken away from as the user expands/contracts nodes.

A useful example would be a file browser. When expanding a folder the listing is fetched and added to the tree. When collapsing the folder listing is removed from the tree.