More efficient toggling of expanded state
When toggling a row the entire DataView was being rendered each time. This PR switches from calling `RenderView` to `RefreshViewRows`. `RefreshViewRows` is much more efficient at displaying the new child rows but won’t call FillInData for the parent row that was toggled because it doesn’t know the data has changed. That means the toggle UI element won’t be updated.

`RefreshViewRows` will update any rows with cache keys that don’t exist. With that in mind, the cache key for each row is now a combination of the node `id` value as well as an internal `@version` property. If the data for a node changes it is marked as dirty and the `@version` will be incremented, thus changing the cache key. When the cache key changes the DataView will update the data for that row.