Add handlers for manipulating tree (#8)
Added the following handlers for manipulating tree nodes:
- AddNode
- MoveNode
- DeleteNode

Fixed an infinite loop bug in `dvNodeAncestorNodes`.

`ToggleNodeIsExpanded` and `ToggleRowIsExpanded` now return the the new `expanded` state in `it`.

The underlying tree structure has changed in order to support this. The tree now has a `root` node at the root which has a `children` node. The first level of nodes are contained in this `["root"]["children"]` array key. This makes the code to manipulate nodes much cleaner.

`dvTree` still supports an array with numerically indexed keys at the root. It will detect this type of array and assign the array to the `["root"]["children"]` key.